+++ ToDo/Service/Services/UserService.cs
@@ -1,8 +1,12 @@
 ﻿using Domain.Entities;
+using Domain.Exceptions;
 using Domain.Interfaces.Repositories;
 using Domain.Interfaces.Services;
 using Domain.Requests.User;
+using Domain.Responses.Register;
 using Domain.Responses.User;
+using Microsoft.AspNetCore.Http;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace Service.Services
 {
@@ -11,17 +15,20 @@ namespace Service.Services
 		private readonly IEncryptService _encryptService;
 		private readonly IUserRepository _userRepository;
 		private readonly ITokenService _tokenService;
-		public UserService(IEncryptService encryptService, IUserRepository userRepository, ITokenService tokenService)
+		private readonly IHttpContextAccessor _httpContextAccessor;
+		public UserService(IEncryptService encryptService, IUserRepository userRepository, ITokenService tokenService, IHttpContextAccessor httpContextAccessor)
 		{
 			_encryptService = encryptService;
 			_userRepository = userRepository;
 			_tokenService = tokenService;
+			_httpContextAccessor = httpContextAccessor;
 		}
 		public async Task<bool> DeleteAsync(Guid id)
 		{
+			var userId = _httpContextAccessor.HttpContext.User.Claims.FirstOrDefault(c => c.Type == JwtRegisteredClaimNames.NameId)?.Value!;
 			var user = await _userRepository.GetAsync(id);
-			if (user == null)
-				return false;
+			if (user == null || !userId.Equals(user.Id.ToString()))
+				throw new DomainException("User not found", 400);
 
 			await _userRepository.DeleteAsync(user);
 			return true;
@@ -40,14 +47,14 @@ namespace Service.Services
 			return await _userRepository.CreateAsync(entity);
 		}
 
-		public async Task<string> LoginAsync(LoginRequest login)
+		public async Task<TokenResponse> LoginAsync(LoginRequest login)
 		{
 			var user = await _userRepository.GetAsync(login.User);
 			var validPassword = _encryptService.CheckHash(login.Password, user.Password ?? "");
 			if (user == null || !validPassword)
-				return "";
+				throw new DomainException("User not found", 400);
 
-			return _tokenService.GenerateToken(user);
+			return new TokenResponse(_tokenService.GenerateToken(user));
 		}
 	}
 }
